+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 1.4.2 ###
+* Drop recursive check on option that failed in several scenarios (#96) (#97)
+* Drop check for base64 encoded strings which did not work properly in al cases (#100)
+* Use WP 5.7 color palette for the UI (#99)
+
 ### 1.4.1 ###
 * Fix some spelling mistakes and correct translations (#85)
 * Fix file name sanitization in manual theme scan causing errors to be not shown in the admin area (#88, #89)